django/pyproject.toml
[project]
dependencies = [
  "celery-redbeat>=2.2.0",
  "celery>=5.4.0",
  "django-cors-headers>=4.4.0",
  "django-extensions>=3.2.3",
  "django-filter>=24.3",
  "django-storages>=1.14.4",
  "django-stubs-ext>=5.0.4",
  "django-stubs>=5.0.4",
  "django-constance>=4.1.2",
  "django==4.2.1",
  "gunicorn>=23.0.0",
  "pillow>=10.4.0",
  "psycopg[binary]>=3.2.1",
  "uvicorn>=0.30.6",
  "django-admin-list-filter-dropdown>=1.0.3",
  "djangorestframework>=3.15.2",
]
description = "Django test app"
name = "backend"
requires-python = ">=3.12"
version = "0.1.0"

[project.optional-dependencies]
dev = [
  "djlint>=1.35.2",
  'ruff>=0.8.1',
]

[tool.ruff]
exclude = [
  ".git",
  ".mypy_cache",
  ".ruff_cache",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "migrations",
  "dist",
  ".venv",
  ".conf",
  "asgi.py",
]
line-length = 120
target-version = "py312"
[tool.ruff.lint]
extend-select = ["I"]
# Exclude a variety of commonly ignored directories.
exclude = [
  ".git",
  ".mypy_cache",
  ".ruff_cache",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "migrations",
  "dist",
  ".venv",
  ".conf",
  "asgi.py",
]
ignore = [
  "ANN",
  "D",
  "PT009",
  "B904",
  "S101",
  "ARG001",
  "ARG002",
  "G004",
  "EM102",
  "TRY003",
  "TRY401",
  "EM101",
  "S311",
  "S107",
  "S105",
  "N802",
  "PTH",
  "RUF012",
  "TD002",
  "TD003",
  "FIX002",
  "FBT003",
  "PD002",
  "COM812",
  "ISC001",
  "FBT001",
  "FBT002",
  "ERA001",
]
select = ["ALL"]

per-file-ignores = {}
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
[tool.ruff.lint.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

[tool.ruff.format]
# Like Black, use double quotes for strings.
indent-style = "space"
line-ending = "auto"
quote-style = "double"
skip-magic-trailing-comma = false

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.coverage.run]
branch = true
omit = [
  ".venv/*",
  "**/migrations/**",
  "*/admin.py",
  "*/apps.py",
  "manage.py",
  "*/settings.py",
  "*/asgi.py",
  "*/wsgi.py",
  "tests.py",
]

[tool.coverage.report]
show_missing = true
skip_empty = true
sort = "-miss"

[tool.uv.sources]
django-erp-framework = { git = "https://github.com/RamezIssac/django-erp-framework.git", rev = "develop" }
django-slick-reporting = { git = "https://github.com/ra-systems/django-slick-reporting.git", rev = "develop" }
